requirements.txt
# Core dependencies
beautifulsoup4>=4.12.2
requests>=2.31.0
streamlit>=1.28.0
crewai>=0.28.0
python-dotenv>=1.0.0

# LLM and AI dependencies
google-generativeai>=0.8.4  # Updated to latest version for Gemini
langchain>=0.0.335
langchain_community>=0.0.1

# Text processing
nltk>=3.8.1
tqdm>=4.66.1

# LiteLLM for Gemini configuration
litellm>=1.0.0
